requirements.txt
# Core MCP (Model Context Protocol) dependencies
mcp>=1.10.0

# Google API dependencies
google-auth>=2.15.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# Google Generative AI (Gemini) dependencies
google-generativeai>=0.8.0

# GitHub API dependencies
PyGithub>=2.0.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP and networking
requests>=2.28.0

# Data handling and utilities
pydantic>=2.0.0

# Async support
anyio>=3.0.0

# Development and testing dependencies (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Additional utilities
tqdm>=4.64.0
typing-extensions>=4.0.0 